operationsservice/requirements.txt
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.1
Flask==3.0.3
Flask-JWT-Extended==4.6.0
idna==3.7
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mysql-connector-python==8.3.0
packaging==24.0
pluggy==1.5.0
PyJWT==2.8.0
pytest==8.2.1
python-dotenv==1.0.1
requests==2.31.0
tomli==2.0.1
urllib3==2.2.1
Werkzeug==3.0.2
